Pretty hacky way of making flymake work with tramp.
This causes flymake to check the file locally, which means (among
other things) imports will be pulled from the local environment so
remote changes to imports will not be reflected.

Ideally flymake would be run on the remote machine, this requires
creating the temporary file on the remote machine though. On a slow
link this can cause emacs to hang while the file is transfered.